pyproject.toml
[tool.poetry]
name = "shutter-api"
version = "0.1.0"
description = "Api For Shutter"
authors = ["Jeremi Girard <jeremigirard@hotmail.com>"]
readme = "README.md"
packages = [{ include = "shutter_api" }]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.3"
flask-mysql = "^1.5.2"
flask-cors = "^3.0.10"
flask-jwt-extended = "^4.4.4"
imagekitio = "^3.0.1"
pycryptodome = "^3.17"

[tool.poetry.scripts]
serve = "shutter_api.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
